refactor: move registry tests out of ArtifactoryRegistryTest
